pyproject.toml
[tool.poetry]
name = "depcos2024"
version = "0.1.0"
description = ""
authors = ["Ratajczak, Najda, Szyc"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
numpy = ">=1.26.2,<1.27.0"
librosa = ">=0.10.1,<0.11.0"
matplotlib = ">=3.8.2,<3.9.0"
torch = ">=2.1.1,<2.2.0"
torchvision = ">=0.16.1,<0.17.0"
scikit-learn = ">=1.3.2,<1.4.0"
pillow = ">=10.1.0,<10.2.0"
torchaudio = ">=2.1.1,<2.2.0"
tqdm = ">=4.66.1,<4.67.0"
gdown = "4.7.3"
wandb = "0.16.2"
python-dotenv = "^1.0.1"
timm = "^1.0.9"
more-itertools = "^10.5.0"
grad-cam = "^1.5.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
